refactor(KlaviyoSwift): relocate geofence dispatch off KlaviyoInternal (MAGE-833)
